requirements.txt
flask
flask-mail
flask-sqlalchemy
mysql-connector-python
werkzeug
pandas
openpyxl
gunicorn